lazily import pyqt + adjust progress handling
