+++ picasso/g5m.py
@@ -22,17 +22,18 @@ import time
 from abc import ABCMeta, abstractmethod
 from concurrent.futures import ProcessPoolExecutor
 from itertools import chain as itchain
-from typing import Literal, Any
+from typing import Literal, TYPE_CHECKING
 
 import numpy as np
 import pandas as pd
 from numba import njit
 from scipy.special import erf
 from sklearn.utils import check_random_state
-from tqdm import tqdm
-from PyQt6 import QtWidgets
 
 from . import lib, zfit, __version__
+
+if TYPE_CHECKING:
+    from PyQt6 import QtWidgets  # only used in type annotations
 
 # default min. number of localizations per molecule
 MIN_LOCS = 10
@@ -2353,8 +2354,7 @@ def _g5m(
     max_locs_per_cluster: int,
     asynch: bool,
     n_steps: int,
-    progress: Any,
-    callback_parent: Any,
+    progress: lib.ProgressType,
     mode: Literal["astigmatism", "spline"] = "astigmatism",
 ) -> tuple[list[pd.DataFrame], list[pd.DataFrame]]:
     """Run G5M with or without multiprocessing. The function returns the
@@ -2377,10 +2377,7 @@ def _g5m(
         # display progress
         while lib.n_futures_done(fs) < n_steps:
             n_done = lib.n_futures_done(fs)
-            if callback_parent != "console":
-                progress.set_value(n_done)
-            else:
-                progress.update(n_done - progress.n)
+            progress.set_value(n_done)
             time.sleep(0.2)
 
         # extract centers from futures
@@ -2421,16 +2418,11 @@ def _g5m(
                 centers.append(centers_)
                 clustered_locs.append(clustered_locs_)
 
-            if callback_parent == "console":
-                progress.update(1)
-            else:
-                progress.set_value(i)
+            progress.set_value(i)
 
-    # close progress widget if present
-    if callback_parent != "console":
-        progress.close()
-    else:
-        progress.update(1)
+    # complete and close the progress tracker
+    progress.set_value(n_steps)
+    progress.close()
 
     return centers, clustered_locs
 
@@ -2576,11 +2568,13 @@ def g5m(
     # determine how many steps are displayed in the progress bar
     n_steps = N_TASKS if asynch else len(np.unique(locs["group"]))
 
-    # initialize the progress bar
-    if callback_parent == "console":
-        progress = tqdm(total=n_steps, desc="Running G5M...")
-    elif callback_parent is None:
-        progress = lib.MockProgress()
+    # initialize the progress bar (uniform ProgressDialog-like
+    # interface, see lib.normalize_progress)
+    if callback_parent is None or callback_parent == "console":
+        progress = lib.normalize_progress(
+            callback_parent, description="Running G5M..."
+        )
+        progress.setMaximum(n_steps)
     else:
         progress = lib.ProgressDialog(
             "Running G5M...", 0, n_steps, callback_parent
@@ -2600,7 +2594,6 @@ def g5m(
         asynch=asynch,
         n_steps=n_steps,
         progress=progress,
-        callback_parent=callback_parent,
         mode=mode,
     )
     # stack centers to form a pd.DataFrame in the format of localizations
